v0.8: Implementing Decimal support, User Preferences (Goal/Rent), Dynamic Calendar navigation, Achievement Badges, and bug fixes

+++ app/src/main/java/com/hisabmate/data/local/entities/DailyRecord.kt
@@ -8,8 +8,8 @@ data class DailyRecord(
     @PrimaryKey
     val date: Long, // Epoch millis for start of day
     
-    val mealsCount: Int = 0,
-    val teasCount: Int = 0,
+    val mealsCount: Double = 0.0,
+    val teasCount: Double = 0.0,
     val moneyAmount: Double = 0.0,
     
     val createdAt: Long = System.currentTimeMillis(),
